+++ src/deps.mts
@@ -1,0 +1,4 @@
+import * as path from 'node:path'
+import { fileURLToPath } from 'node:url'
+export const __filename = fileURLToPath(import.meta.url)
+export const __dirname = path.dirname(__filename)
